Trait ovmi::executor::ExternalCall

pub trait ExternalCall {
    type Address: MaybeAddress;
    type Hash: MaybeHash;
    type Hashing: Hasher<Out = Self::Hash>;

    const NotPredicate: Self::Address;
    const AndPredicate: Self::Address;

    fn ext_call(
        &mut self, 
        to: &Self::Address, 
        input_data: PredicateCallInputs<Self::Address>
    ) -> ExecResult<Self::Address>;

    fn ext_caller(&self) -> Self::Address;

    fn ext_address(&self) -> Self::Address;

    fn ext_is_stored(
        &mut self, 
        address: &Self::Address, 
        key: &[u8], 
        value: &[u8]
    ) -> bool;

    fn ext_is_decided(&self, property: &PropertyOf<Self>) -> bool;

    fn ext_get_property_id(&self, property: &PropertyOf<Self>) -> Self::Hash;

    fn ext_set_predicate_decision(
        &self, 
        game_id: Self::Hash, 
        decision: bool
    ) -> ExecResult<Self::Address>;

    fn hash_of<S: Encode>(s: &S) -> Self::Hash { ... }
}

Associated Types

type Address: MaybeAddress

type Hash: MaybeHash

type Hashing: Hasher<Out = Self::Hash>

Loading content...
Associated Constants

const NotPredicate: Self::Address

const AndPredicate: Self::Address

Loading content...
Required methods

fn ext_call(
&mut self,
to: &Self::Address,
input_data: PredicateCallInputs<Self::Address>
) -> ExecResult<Self::Address>

Call (other predicate) into the specified account.

fn ext_caller(&self) -> Self::Address

Returns a reference to the account id of the caller.

fn ext_address(&self) -> Self::Address

Returns a reference to the account id of the current contract.

fn ext_is_stored(
&mut self,
address: &Self::Address,
key: &[u8],
value: &[u8]
) -> bool

Notes a call other storage. Only return true or false. CommitmentAddress(special) isCommitment(address) -> Commitment is_stored_predicate(&mut self, address, key, value);? ref: https://github.com/cryptoeconomicslab/ovm-contracts/blob/master/contracts/Predicate/Atomic/IsStoredPredicate.sol

fn ext_is_decided(&self, property: &PropertyOf<Self>) -> bool

is_decided function of UniversalAdjudication in OVM module.

fn ext_get_property_id(&self, property: &PropertyOf<Self>) -> Self::Hash

get_property_id function of UniversalAdjudication in OVM module.

fn ext_set_predicate_decision(
&self,
game_id: Self::Hash,
decision: bool
) -> ExecResult<Self::Address>

set_predicate_decision function of UniversalAdjudication in OVM module.

Loading content...
Provided methods

fn hash_of<S: Encode>(s: &S) -> Self::Hash

Produce the hash of some codec-encodable value.

Loading content...