Module ovmi::predicates

Executable Predicates.

Executable Predicates instanced from Compiled Predicates and Atomic Predicates.

Structs

AndPredicate

CompiledExecutable

NotPredicate

Property
Property stands for dispute logic and we can claim every Properties to Adjudicator Contract. Property has its predicate address and array of input.

Enums

AtomicPredicateCallInputs

BaseAtomicPredicateCallInputs

CompiledPredicateCallInputs

DeciableExecutable

DecidablePredicateCallInputs

LogicalConnectiveCallInputs

LogicalConnectiveExecutable

PredicateCallInputs

Traits

AtomicPredicateInterface

BaseAtomicPredicateInterface

CompiledPredicateInterface

DecidablePredicateInterface

LogicalConnectiveInterface

UniversalAdjudication

Utils